movie-api/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.21
uWSGI==2.0.19.1
MarkupSafe==2.0.1
flask_cors